feat: add vscode extension

+++ src/generator/generator.ts
@@ -13,9 +13,9 @@ function genPrefix(first: string, relativePath: string, ingoreDefaultMd: boolean
   res.push(
     ...relativePath
       .split(path.sep)
-      .filter(w => w.length > 0)
-      .map(w => w.endsWith('.md') ? w.substr(0, w.length - 3) : w)
-      .filter(w => w.length > 0)
+      .filter((w) => w.length > 0)
+      .map((w) => (w.endsWith('.md') ? w.substr(0, w.length - 3) : w))
+      .filter((w) => w.length > 0),
   );
   if (ingoreDefaultMd && res.length > 1 && res.includes('default')) {
     res.pop();
@@ -25,10 +25,13 @@ function genPrefix(first: string, relativePath: string, ingoreDefaultMd: boolean
 
 function cleanObject(obj: { [key: string]: Snippet }) {
   const res: { [key: string]: Snippet } = {};
-  Object.keys(obj).forEach(key => {
+  Object.keys(obj).forEach((key) => {
     const { prefix, body, description, scope } = obj[key];
     res[key] = {
-      prefix, body, description, scope
+      prefix,
+      body,
+      description,
+      scope,
     } as Snippet;
   });
   return res;
@@ -37,38 +40,35 @@ function cleanObject(obj: { [key: string]: Snippet }) {
 export function generator(config?: ConfigSchema, options?: GenerateOptions) {
   const cog = {
     ...DEFAULT_CONFIG,
-    ...config
+    ...config,
   } as ConfigSchema;
-  const sourcePaths: Array<{ rootPath: string, fullPath: string }> = [];
+  const sourcePaths: { rootPath: string; fullPath: string }[] = [];
   (typeof cog.sourceRoot === 'string' ? [cog.sourceRoot] : cog.sourceRoot)
-    .map(p => path.resolve(process.cwd(), p))
-    .filter(p => fs.existsSync(p))
-    .forEach(rootPath => {
-      klaw(
-        rootPath,
-        {
-          nodir: true,
-          traverseAll: true,
-          filter: (item: klaw.Item) => item.path.endsWith('.md')
-        } as any)
-        .map(item => item.path)
-        .forEach(fullPath => {
+    .map((p) => path.resolve(process.cwd(), p))
+    .filter((p) => fs.existsSync(p))
+    .forEach((rootPath) => {
+      klaw(rootPath, {
+        nodir: true,
+        traverseAll: true,
+        filter: (item: klaw.Item) => item.path.endsWith('.md'),
+      } as any)
+        .map((item) => item.path)
+        .forEach((fullPath) => {
           sourcePaths.push({ rootPath, fullPath });
         });
     });
 
   let res: { [key: string]: Snippet } = {};
-  sourcePaths.forEach(sourceItem => {
+  sourcePaths.forEach((sourceItem) => {
     const relativePath = path.relative(sourceItem.rootPath, sourceItem.fullPath);
     const md = fs.readFileSync(sourceItem.fullPath).toString('utf8');
     try {
       let item = parse(md, sourceItem.fullPath, cog);
       const keys = genPrefix(cog.prefix, relativePath, cog.ingoreDefaultMd);
-      item.prefix = keys.join(cog.separator);
+      item.prefix = item.prefix || keys.join(cog.separator);
       if (options && options.itemFinished) {
         item = options.itemFinished(item);
       }
-
       res[keys.join('_')] = item;
     } catch (err) {
       console.log(`>> File ${relativePath} parse error: ${chalk.red(err)}`);
@@ -88,11 +88,13 @@ export function generator(config?: ConfigSchema, options?: GenerateOptions) {
     res = options.finished(res);
   }
 
-  const content = JSON.stringify(res, null, 2);
-  const saveFile = path.resolve(process.cwd(), cog.outFile);
-  if (fs.existsSync(saveFile)) {
-    fs.unlinkSync(saveFile);
+  if (cog.outFile != null) {
+    const content = JSON.stringify(res, null, 2);
+    const saveFile = path.resolve(process.cwd(), cog.outFile);
+    if (fs.existsSync(saveFile)) {
+      fs.unlinkSync(saveFile);
+    }
+    fs.writeFileSync(saveFile, content);
   }
-  fs.writeFileSync(saveFile, content);
   return res;
 }
